+++ src/eg3dproj.c
@@ -47,7 +47,7 @@ void projectObjects(int heading, int rangeGate, long worldX, long worldY, long w
     worldX = g_proj3d.x;
     worldY = g_proj3d.y;
     worldZ = g_proj3d.z;
-    dirSector = (unsigned)(-heading + 0x1000) >> 13;
+    dirSector = (uint16)(-heading + 0x1000) >> 13;
     g_curLod = (g_detailLevel != 0) ? 4 : 3;
     goto outer_test;
     do {
@@ -73,8 +73,8 @@ outer_test:
                 if (sampleIdx == 15) {
                     break;
                 }
-                gridX = *(const int *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)dirSector);
-                gridY = *(const int *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)((dirSector + 2) & 7));
+                gridX = *(const int16 *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)dirSector);
+                gridY = *(const int16 *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)((dirSector + 2) & 7));
                 g_objLocalX = fracX - (gridX << 12) - 0x800;
                 g_objLocalY = fracY - (gridY << 12) - 0x800;
                 g_objRenderMode = 7;
@@ -92,8 +92,8 @@ outer_test:
                     gridX = g_neighborSampling.gridX[sampleIdx];
                     gridY = g_neighborSampling.gridY[sampleIdx];
                 } else {
-                    gridX = *(const int *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)dirSector);
-                    gridY = *(const int *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)((dirSector + 2) & 7));
+                    gridX = *(const int16 *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)dirSector);
+                    gridY = *(const int16 *)((const char *)g_dirGridOffsets + sampleIdx * 2 + (unsigned)18 * (unsigned)((dirSector + 2) & 7));
                 }
                 g_objLocalX = fracX - (gridX << 12) - 0x800;
                 g_objLocalY = fracY - (gridY << 12) - 0x800;
